src/requirements.txt
flask
python-escpos
pillow
